+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
-## 0.2.0 unpublished
+## 0.2.1 unpublished
+- Add `FormPage` element `settingRender` props, Custom Render Drawer Form Extension
+
+## 0.2.0
 
 ### Feature
 - Add basic rich text editor using tiptap
-- Add form `actionRender` props
+- Add `FormPage` element `actionRender` props
 
 ## 0.0.1
 - init project